code commit for controller cheanges

+++ app/Http/Controllers/ExpenditureController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Expenditure;
 use App\Models\Item;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
 class ExpenditureController extends Controller
@@ -26,6 +27,8 @@ class ExpenditureController extends Controller
      */
     public function create(Request $request)
     {
+
+
         //
         $ex = new Expenditure();
         $ex->expenditure_name = $request['expenditure_name'];
@@ -55,11 +58,20 @@ class ExpenditureController extends Controller
      */
     public function show(string $id)
     {
+        $user = Auth::user();
+        // $ex = Expenditure::findOrFail($id);
+
+
         $ex = Expenditure::find($id);
-        $itemEx = DB::table('expenditures')
-            ->join('items', 'items.id', '=', 'expenditures.item_id')
-            ->get();
-        return view('expenditure.edit', compact('ex', 'itemEx'));
+        if ($user->id === $ex->user_id || Auth::user()->id == '1') {
+
+            $itemEx = DB::table('expenditures')
+                ->join('items', 'items.id', '=', 'expenditures.item_id')
+                ->get();
+            return view('expenditure.edit', compact('ex', 'itemEx'));
+        } else {
+            return \response()->json("can not show");
+        }
     }
 
     /**
@@ -75,18 +87,27 @@ class ExpenditureController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $ex =  Expenditure::find($id);
-        $ex->expenditure_name = $request['expenditure_name'];
-        $ex->expenditure_type = $request['expenditure_type'];
-        $ex->expenditure_amount = $request['expenditure_amount'];
-        $ex->item_id = $request['item_id'];
-        $ex->user_id = auth()->user()->id;
-        try {
-            $ex->save();
-            return redirect()->route('expenditure.index');
-            //code...
-        } catch (\Throwable $th) {
-            throw $th;
+
+        $user = Auth::user();
+        // dd($user);
+        $ex = Expenditure::findOrFail($id);
+        if ($user->id == $ex->user_id || Auth::user()->id == '1') {
+
+            $ex =  Expenditure::find($id);
+            $ex->expenditure_name = $request['expenditure_name'];
+            $ex->expenditure_type = $request['expenditure_type'];
+            $ex->expenditure_amount = $request['expenditure_amount'];
+            $ex->item_id = $request['item_id'];
+            $ex->user_id = auth()->user()->id;
+            try {
+                $ex->save();
+                return redirect()->route('expenditure.index');
+                //code...
+            } catch (\Throwable $th) {
+                throw $th;
+            }
+        } else {
+            return \response()->json("can not Updated");
         }
     }
 
@@ -95,8 +116,17 @@ class ExpenditureController extends Controller
      */
     public function destroy(string $id)
     {
+        $user = Auth::user();
+        // dd($user);
+        $ex = Expenditure::findOrFail($id);
+        if ($user->id == $ex->user_id || Auth::user()->id == '1') {
+
+            Expenditure::find($id)->delete();
+            return redirect()->route('expenditure.index');
+        } else {
+            return \response()->json("can not Deleted");
+        }
+
         //
-        Expenditure::find($id)->delete();
-        return redirect()->route('expenditure.index');
     }
 }
